qcom-multimedia-proprietary-image: add image to showcase proprietary features
This qcom-multimedia-proprietary-image, showcases capabilities that are either
missing or not fully supported in upstream, particularly in areas such as
camera, display, GPU, video and other proprietary multimedia components.

It inherits qcom-multimedia-image, which provides a baseline for upstream
functionality, and extends or modifies those features to deliver a tailored
experience with proprietary enhancements.

Signed-off-by: Viswanath Kraleti <viswanath.kraleti@oss.qualcomm.com>